REF: refactor describe.py into the methods folder
