+++ pandas/core/methods/describe.py
